// src/app/mill_management_system/model/Member/member-master.model.ts
export class MemberMaster {
    memberName: string | any;
    stateCode: string | any;
    taluka: string | any;
    contactNo: string | any;
    address: string | any;
    memberType: string | any;
    districtCode: string | any;
    pinCode:String|any;
    emailId:String|any;
    userId:String|any;
    
}