// src/app/mill_management_system/model/User.ts
export class User {

    userId: string|any;
    orgId: string|any;
    userName: string|any;
    userCategory: string|any;
    userMobile: string|any;
    userEmail: string|any;
    userAddress: string|any;
    userFirstLogin: string|any;
    userIsActive: string|any;
    orgUserScope: string|any;
}